Restructured code to allow defining mcp tools in multiple python modules. Modules can be dynamically loaded by requirement using a config file to match Couchbase multimodel architecture.
