+++ src/mcp_config.json
@@ -1,0 +1,3 @@
+{
+    "tools":["tools"]
+}
